requirements/base.txt
pytz==2022.1  # https://github.com/stub42/pytz
python-slugify==6.1.2  # https://github.com/un33k/python-slugify
Pillow==9.1.1  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.2.0  # https://github.com/evansd/whitenoise
redis==4.3.3  # https://github.com/redis/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==3.2.13  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.9.0  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.51.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis

django-debug-toolbar~=3.4.0
pytest~=7.1.2
django-filter~=22.1
